fix documentation
