+++ be1-go/configServer1.json
@@ -8,5 +8,6 @@
     "client-port"                          : 9000,
     "server-port"                          : 9001,
     "auth-port"                            : 9100,
-    "other-servers": []
-}
+    "other-servers": [
+        "localhost:9003", "localhost:9005", "localhost:9007"
+    ]}
